requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
scikit-learn==0.21.3
scikit-surprise==1.0.6
scipy==1.3.1
six==1.12.0
Werkzeug==0.16.0